#191 refactor: User 패키지 안에 있던 Member패키지를 외부로 이동
