+++ src/main/java/store/novabook/store/cart/dto/GetCartResponse.java
@@ -1,7 +1,7 @@
 package store.novabook.store.cart.dto;
 
 import lombok.Builder;
-import store.novabook.store.user.member.entity.Member;
+import store.novabook.store.member.entity.Member;
 
 @Builder
 public record GetCartResponse(
