+++ src/main/java/store/novabook/store/common/config/JWTFilter.java
@@ -17,9 +17,9 @@ import store.novabook.store.common.security.dto.GetNewTokenRequest;
 import store.novabook.store.common.security.dto.GetNewTokenResponse;
 import store.novabook.store.common.security.entity.Users;
 import store.novabook.store.common.security.service.NewTokenClient;
-import store.novabook.store.user.member.MemberClient;
-import store.novabook.store.user.member.dto.GetMembersUUIDRequest;
-import store.novabook.store.user.member.dto.GetMembersUUIDResponse;
+import store.novabook.store.member.MemberClient;
+import store.novabook.store.member.dto.GetMembersUUIDRequest;
+import store.novabook.store.member.dto.GetMembersUUIDResponse;
 
 @Slf4j
 public class JWTFilter extends OncePerRequestFilter {
