+++ src/main/java/store/novabook/store/member/MemberClient.java
@@ -1,11 +1,11 @@
-package store.novabook.store.user.member;
+package store.novabook.store.member;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import store.novabook.store.user.member.dto.GetMembersUUIDRequest;
-import store.novabook.store.user.member.dto.GetMembersUUIDResponse;
+import store.novabook.store.member.dto.GetMembersUUIDRequest;
+import store.novabook.store.member.dto.GetMembersUUIDResponse;
 
 @FeignClient(name = "memberClient", url = "http://localhost:9777/auth/uuid")
 public interface MemberClient {
