+++ src/main/java/store/novabook/store/member/dto/GetMemberGradeResponse.java
@@ -1,7 +1,7 @@
-package store.novabook.store.user.member.dto;
+package store.novabook.store.member.dto;
 
 import lombok.Builder;
-import store.novabook.store.user.member.entity.MemberGradeHistory;
+import store.novabook.store.member.entity.MemberGradeHistory;
 
 @Builder
 public record GetMemberGradeResponse(
