+++ src/main/java/store/novabook/store/member/dto/GetMembersUUIDResponse.java
@@ -1,4 +1,4 @@
-package store.novabook.store.user.member.dto;
+package store.novabook.store.member.dto;
 
 public record GetMembersUUIDResponse(
 	String usersId
