+++ src/main/java/store/novabook/store/member/dto/LoginMemberRequest.java
@@ -1,4 +1,4 @@
-package store.novabook.store.user.member.dto;
+package store.novabook.store.member.dto;
 
 public record LoginMemberRequest(String loginId, String loginPassword) {
 }
