+++ src/main/java/store/novabook/store/member/entity/Member.java
@@ -1,4 +1,4 @@
-package store.novabook.store.user.member.entity;
+package store.novabook.store.member.entity;
 
 import java.time.LocalDateTime;
 
@@ -18,7 +18,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import store.novabook.store.user.member.dto.CreateMemberRequest;
+import store.novabook.store.member.dto.CreateMemberRequest;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
