+++ src/main/java/store/novabook/store/member/repository/MemberAddressRepository.java
@@ -1,11 +1,11 @@
-package store.novabook.store.user.member.repository;
+package store.novabook.store.member.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import store.novabook.store.user.member.entity.Member;
-import store.novabook.store.user.member.entity.MemberAddress;
+import store.novabook.store.member.entity.Member;
+import store.novabook.store.member.entity.MemberAddress;
 
 public interface MemberAddressRepository extends JpaRepository<MemberAddress, Long> {
 	Integer countByMember(Member member);
