+++ src/main/java/store/novabook/store/member/repository/MemberCouponRepository.java
@@ -1,8 +1,8 @@
-package store.novabook.store.user.member.repository;
+package store.novabook.store.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import store.novabook.store.user.member.entity.MemberCoupon;
+import store.novabook.store.member.entity.MemberCoupon;
 
 public interface MemberCouponRepository extends JpaRepository<MemberCoupon, Long> {
 }
