+++ src/main/java/store/novabook/store/member/repository/MemberGradeHistoryRepository.java
@@ -1,10 +1,10 @@
-package store.novabook.store.user.member.repository;
+package store.novabook.store.member.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import store.novabook.store.user.member.entity.MemberGradeHistory;
+import store.novabook.store.member.entity.MemberGradeHistory;
 
 public interface MemberGradeHistoryRepository extends JpaRepository<MemberGradeHistory, Long> {
 	Optional<MemberGradeHistory> findByMemberId(Long memberId);
