+++ src/main/java/store/novabook/store/member/repository/StreetAddressRepository.java
@@ -1,8 +1,8 @@
-package store.novabook.store.user.member.repository;
+package store.novabook.store.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import store.novabook.store.user.member.entity.StreetAddress;
+import store.novabook.store.member.entity.StreetAddress;
 
 public interface StreetAddressRepository extends JpaRepository<StreetAddress, Long> {
