+++ src/main/java/store/novabook/store/member/service/MemberCouponService.java
@@ -1,4 +1,4 @@
-package store.novabook.store.user.member.service;
+package store.novabook.store.member.service;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Service;
@@ -10,13 +10,14 @@ import store.novabook.store.common.adatper.CouponAdapter;
 import store.novabook.store.common.adatper.dto.CreateCouponRequest;
 import store.novabook.store.common.adatper.dto.CreateCouponResponse;
 import store.novabook.store.common.exception.EntityNotFoundException;
+import store.novabook.store.member.dto.CreateMemberCouponRequest;
+import store.novabook.store.member.dto.CreateMemberCouponResponse;
+import store.novabook.store.member.entity.Member;
+import store.novabook.store.member.entity.MemberCoupon;
+import store.novabook.store.member.repository.MemberCouponRepository;
+import store.novabook.store.member.repository.MemberRepository;
 import store.novabook.store.message.CouponCreatedMessage;
-import store.novabook.store.user.member.dto.CreateMemberCouponRequest;
-import store.novabook.store.user.member.dto.CreateMemberCouponResponse;
-import store.novabook.store.user.member.entity.Member;
-import store.novabook.store.user.member.entity.MemberCoupon;
-import store.novabook.store.user.member.repository.MemberCouponRepository;
-import store.novabook.store.user.member.repository.MemberRepository;
+
 
 @RequiredArgsConstructor
 @Service
