+++ src/main/java/store/novabook/store/member/service/MemberGradeHistoryService.java
@@ -1,13 +1,13 @@
-package store.novabook.store.user.member.service;
+package store.novabook.store.member.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import store.novabook.store.common.exception.EntityNotFoundException;
-import store.novabook.store.user.member.dto.GetMemberGradeResponse;
-import store.novabook.store.user.member.entity.MemberGradeHistory;
-import store.novabook.store.user.member.repository.MemberGradeHistoryRepository;
+import store.novabook.store.member.dto.GetMemberGradeResponse;
+import store.novabook.store.member.entity.MemberGradeHistory;
+import store.novabook.store.member.repository.MemberGradeHistoryRepository;
 
 @Service
 @RequiredArgsConstructor
