+++ src/main/java/store/novabook/store/point/entity/PointHistory.java
@@ -19,8 +19,8 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import store.novabook.store.member.entity.Member;
 import store.novabook.store.orders.entity.Orders;
-import store.novabook.store.user.member.entity.Member;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
