+++ src/main/java/store/novabook/store/point/service/PointHistoryService.java
@@ -9,6 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import store.novabook.store.common.exception.EntityNotFoundException;
+import store.novabook.store.member.entity.Member;
+import store.novabook.store.member.repository.MemberRepository;
 import store.novabook.store.orders.entity.Orders;
 import store.novabook.store.orders.repository.OrdersRepository;
 import store.novabook.store.point.dto.CreatePointHistoryRequest;
@@ -17,8 +19,6 @@ import store.novabook.store.point.entity.PointHistory;
 import store.novabook.store.point.entity.PointPolicy;
 import store.novabook.store.point.repository.PointHistoryRepository;
 import store.novabook.store.point.repository.PointPolicyRepository;
-import store.novabook.store.user.member.entity.Member;
-import store.novabook.store.user.member.repository.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
